Fix freeze constraints to include flags of local packages
We deliberately exclude version constraints of packages that are local
to the project (since their versions are fixed to be what they are
anyway, and it'd just make it harder to bump versions etc). However the
flag constraints of local packages were also being ignored, which leads
to "slushy" frozen files.

We saw this in practice with leksah where the flag gtk-3.18 was varying
between different users and causing great confusion.