SetupWrapper: split setupWrapper into getSetup and runSetup
The two phases of setupWrapper (configuration and execution) are
separated into the getSetup and runSetup functions. This allows us to
skip configuration if the setup method is known; this is the case when
running multiple commands in sequence or in the new-build
system. Splitting the phases also allows us to choose which _command_ to
run based on the Cabal version in use; setupWrapper demands a fixed
choice of command and only allows the command _flags_ to vary depending
on the version.