Print selected configure flags in install plan.
When debugging #3827, I wished I could see what the profiling
status of all packages in our install plan was.  This patch
outputs this information, by printing the *configure flag*
we would have passed to configure.

In an ideal world, we'd print exactly the configure flags
we actually pass in practice.  But this is not so great because
the flag lists tend to be really long.  So we pare it down
in the following ways:

1. Not all flags are output: only a selected few.  In this
   patch, it's just enabling/disabling profiling flags, but
   we could add more.

2. If we have Flag False for profiling, we will actually
   explicitly pass --disable-profiling to the configure
   script.  But it's not so good to output that to the
   user.  So 'nubFlag' lets us drop the flag, if omitting
   it would have the same effect as specifying it.
   Figuring that out is a little tricky though.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>