Update hackage index before running integration tests
Now that we default to the secure repo, we need to run cabal update the
first time before running cabal (new-)build/etc, which is essentially
what the integration tests do. The travis and appveyor scripts already
did a cabal update at the beginning, but this one was using the
boostrap cabal and currently that version does not use the secure repo
format by default.