Add a sanity unit test for dir mtime behaviour
We rely on posix & ntfs directory mtime semantics, which is that a dir
mtime changes if the set of names in the dir changes.

We rely on this both for the meaning of monitorDirectory but also we
rely on it as an optimisation within the glob checking code.

So this test should always pass on posix-compliant file systems.
Apparently it may not hold on FAT file systems.