Add D.Backpack module with IndefUnitId and IndefModule types
This module defines the key data types 'IndefUnitId' and
'IndefModule', which represent components which are partially
instantiated with holes in them.  The intent is that they
can be substituted over (Distribution.Backpack.ModSubst),
and once all the holes are filled they can be improved into
proper 'UnitId's, which can then be installed.

Also add Distribution.Util.Base62 module containing an
implementation of base-62 encoding, which we use when
computing hashes for fully instantiated unit ids.