Use the new Backpack code for the configure step
Many of the configure functions were factored out and moved to the new
Backpack modules. The new configure action makes a call to
Distribution.Backpack.Configure to setup the ComponentLocalBuildInfo.

Also add an @--instantiate-with@ flag to ./Setup configure, so that
cabal-install can specify how it wants to instantiate a package that
it is building.

Also do the minimal necessary adjustments in cabal-install.