Reduce path lengths for temp package src and build dirs
This is for paths under $dist/tmp/ where we unpack package source
tarballs and build them. This is for packages destined for the store.
The names of these tmp dirs don't matter since they're guaranteed to be
unique, so we only need short identifiers. So stop using full package
names, and put the build artifacts in their own dir, rather than using a
dist dir under the src dir.