New infrastructure for collecting info on available targets
This is one major part of a new implementation of translating from high
level user intentions of what to build into the detail of what specific
components to build.

This part collects and summarises information on what component are
available as targets. Each available target one has enough status info
to determine if we can possibly select the target or if not why not.

This information will be used by each command (build, repl, test etc) to
select the specific targets that a user intention refers to (and to
provide them enough information to be able to produce reasonable error
messages).

Collecting and summarising this info is unfortunately non-trivial
because as part of install plan elaboration we omit components that
cannot be built but these are still components that users could try to
refer to, so we have to reconstruct the info about the omitted
components. It's plausible that we may want to revise the elaborated
install plan representation to make this easier.